+++ source/client/app/AssetFile.hpp
@@ -8,12 +8,12 @@ Minecraft: Pocket Edition - Decompilation Project
 
 #pragma once
 
-#include <cstdint>
+#include <stdint.h>
 
 struct AssetFile {
-    ssize_t size;
+    int64_t size;
     unsigned char *data;
 
     AssetFile(): size(-1), data(nullptr) {}
-    AssetFile(ssize_t size, unsigned char *data): size(size), data(data) {}
+    AssetFile(int64_t size, unsigned char *data): size(size), data(data) {}
 };
